Move Tooltip outside of Draftail package
