+++ examples/components/Tooltip.scss
