+++ examples/components/Tooltip.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
+
 import Tooltip from '../components/Tooltip';
 
 const mockProps = {
