+++ lib/index.js
@@ -1,7 +1,6 @@
 import DraftailEditor from './components/DraftailEditor';
 import Icon from './components/Icon';
 import Portal from './components/Portal';
-import Tooltip from './components/Tooltip';
 import DraftUtils from './api/DraftUtils';
 import { BLOCK_TYPE, ENTITY_TYPE, INLINE_STYLE } from './api/constants';
 
@@ -12,12 +11,4 @@ import { BLOCK_TYPE, ENTITY_TYPE, INLINE_STYLE } from './api/constants';
 
 export default DraftailEditor;
 
-export {
-    DraftUtils,
-    BLOCK_TYPE,
-    ENTITY_TYPE,
-    INLINE_STYLE,
-    Icon,
-    Portal,
-    Tooltip,
-};
+export { DraftUtils, BLOCK_TYPE, ENTITY_TYPE, INLINE_STYLE, Icon, Portal };
