+++ lib/index.scss
@@ -44,4 +44,3 @@ $toolbar-tooltip-z-index: $toolbar-z-index + 1;
 @import './components/ToolbarGroup';
 @import './components/ToolbarButton';
 @import './components/ToolbarTooltip';
-@import './components/Tooltip';
